requirements.txt
pandas==2.1.3
openai==1.3.0
selenium==4.15.2
lxml==4.9.3
python-dotenv==1.0.0
pytest